% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inputData.R
\name{read_inputData}
\alias{read_inputData}
\title{read_inputData}
\usage{
read_inputData(file, nheader, nvars, NAval = -999)
}
\arguments{
\item{file}{Character, pathname to input data file}

\item{nheader}{Integer, number of header lines, usually containing created by and station metadata, in the file}

\item{nvars}{Integer, number of unique variables (ex. runoff, tmin, tmax; nvars=3) in the data file}

\item{NAval}{Integer, value of missing data in statvar/data file (usually -999)}
}
\value{
A dataframe containing the input data file with varialbe_stationNumber column titles
}
\description{
Function to read the input data file used in a PRMS simulation
}
\details{
Variables and station numbers of the data file are determined by the \code{variable numberOfStations}
which is the format read by PRMS and setup by the user prior to a simulation. This function will assign each
timeseries the column title of variable_stationNumber in the order that they are listed.
}
\examples{
exampleData <- read_inputData(file = "data/rghw_ALL.data")
}
\seealso{
\code{\link[prmsTools]{calc_WYstats}},
}
