% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nhruOut.R
\name{read_nhruOut}
\alias{read_nhruOut}
\title{read_nhruOut}
\usage{
read_nhruOut(nhruOutBaseFileName = "", nhruOutVar_names, nhru)
}
\arguments{
\item{nhruOutBaseFileName}{Character, base file name for nhru summary files specified in the PRMS control file. Note
this should also contain the pathname to Output folder containing nhru summary .csv files.}

\item{nhruOutVar_names}{Character vector of variable names to read, based on variables specified in the PRMS control file.}

\item{nhru}{Integer, number of HRUs in PRMS model}
}
\value{
A list of dataframes containing the nhru summary output. List elements are each variable specified in nhruOutVar_names
}
\description{
Function to read the nhru summary files output by PRMS
}
\details{
A function for reading in nhru summary output files. Note that in the PRMS control file, nhruOutON_OFF must
be set to 1 for .csv files to be output.
}
\examples{
exampleData2 <- read_nhruOut(nhruOutBaseFileName = "data/rghw_Daymet_nhru_summary_",
                             nhruOutVar_names = c("hru_rain", "hru_snow"),
                             nhru = 28)
rain <- exampleData2$hru_rain
}
