% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Statvar.R
\name{read_Statvar}
\alias{read_Statvar}
\title{read_Statvar}
\usage{
read_Statvar(file, NAval = -999)
}
\arguments{
\item{file}{Character, pathname to statvar file}

\item{NAval}{Integer, value of missing data in statvar/data file (usually -999)}
}
\value{
A dataframe containing the statvar output with proper column titles
}
\description{
Function to read the statvar file output by PRMS
}
\details{
Variables and index numbers of the statvar file are determined by the user in the PRMS control file
prior to simulation. This function reads the statvar header with variables and index number, and outputs a \code{dataframe}
with column titles corresponding to the statvar elements
}
\examples{
exampleData <- read_Statvar(file = data/rghw_Daymet.statvar)
}
\seealso{
\code{\link[prmsTools]{calc_WYstats}},
}
