% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_WYstats.R
\name{calc_WYstats}
\alias{calc_WYstats}
\title{calc_WYstats}
\usage{
calc_WYstats(data, runoffIndex, segOutflowIndex, writeWYstats = FALSE,
  outFile = NULL)
}
\arguments{
\item{data}{A \code{dataframe} output from \code{read_Statvar}}

\item{runoffIndex}{Integer, index number of observed gage flow in PRMS input data file}

\item{segOutflowIndex}{Integer, index number of segment corresponding to observed gage location}

\item{writeWYstats}{Logical, if \code{TRUE} write .csv file of WY stats.}

\item{outFile}{Character, pathname and filename of .csv file if \code{writeWYstats = TRUE}}
}
\value{
A dataframe containing goodness-of-fit statistics by water year
}
\description{
Function to calculate NSE, RMSE, and VE goodness-of-fit statistics by WY
}
\details{
uses documentation in \code{hydroGOF} package for calculating the following statistics by WY:

NSE: Nash-Sutcliffe Efficiency

RMSE: Root Mean Square Error

VE: Volumetric Efficiency

See \code{hydroGOF} package for more documentation and references
}
\examples{
\dontrun{
data("exampleData",package="prmsTools")
x <- exampleData
basinStats <- calc_WYstats(data = x, runoffIndex = 6, segOutflowIndex = 9)}
}
\seealso{
\code{\link[prmsTools]{read_Statvar}}, \code{\link[hydroGOF]{NSE}},
\code{\link[hydroGOF]{rmse}}, \code{\link[hydroGOF]{VE}}
}
